fix: address code review feedback
Three review issues:

1. The hint's no-args listing row was reverted to "list available MCP
   servers" for self-consistency in the prior round, but that left it
   factually wrong: ``bin/tool``'s ``_list_surface()`` prints both a
   ``Built-ins:`` and an ``MCP servers:`` section unconditionally. Widen
   to "list reachable tools (built-ins + MCP servers)" so the agent's
   expectation matches the binary's output. The hint stays MCP-scoped
   in its grammar rows because the model-tool path is the documented
   built-in interface; the closing sentence already steers built-in
   callers there.

2. Pin the new wording with
   ``test_listing_line_acknowledges_both_builtins_and_mcp_servers`` so
   the same drift can't recur silently.

3. Dedupe the ``tool_module`` fixture: lift it out of
   ``test_tool_cli_help.py`` and ``test_tool_cli_defaults.py`` into a
   shared ``tests/unit/sandbox/conftest.py``. The two files held
   verbatim copies of the ``importlib.SourceFileLoader`` dance for
   loading the extensionless ``bin/tool`` script; conftest discovery
   wires both up automatically.

Mutation-checked: temporarily narrowing the listing row back to "list
available MCP servers" makes the new test fail.
