+++ tests/unit/sandbox/test_tool_cli_help.py
@@ -11,26 +11,9 @@ documentation and parser. These tests lock in the positional form.
 
 from __future__ import annotations
 
-import importlib.machinery
-import importlib.util
-from pathlib import Path
 from types import ModuleType
 
 import pytest
-
-
-@pytest.fixture(scope="module")
-def tool_module() -> ModuleType:
-    """Load ``bin/tool`` as a python module (same pattern as
-    ``test_tool_cli_defaults.py`` — script has no ``.py`` extension)."""
-    repo_root = Path(__file__).parents[3]
-    script_path = repo_root / "bin" / "tool"
-    loader = importlib.machinery.SourceFileLoader("tool_cli", str(script_path))
-    spec = importlib.util.spec_from_file_location("tool_cli", script_path, loader=loader)
-    assert spec is not None and spec.loader is not None
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    return module
 
 
 class TestBuiltinHelpInvokeLine:
